refactor: MonthInformation 구조체 파일로 분리
